simplified; renamed to `_grpct_utils`
